wrap _populate() in try/finally in case of signal
The _populate() method sets a flag to prevent infinite recursion in case a urlconf includes itself. The flag is a threadlocal to avoid race conditions where one thread sets the flag and another checks it, then proceeds to access data that's supposed to be populated (e.g. `_reverse_dict`) but isn't yet.

The potential still exists for a thread to set the threadlocal, then be interrupted by a signal such as SIGALRM ([*ahem*]) and raise before resetting the threadlocal flag. In this scenario, subsequent calls to `_populate()` in the same thread will short-circuit erroneously.

This change handles this potential case by resetting the threadlocal in a `finally` block, which will run before the signal handler.

[1]: https://bitbucket.org/evzijst/interruptingcow